requirements.txt
# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

# LangChain and AI Models
langgraph>=0.5.4
langchain-community>=0.3.9
langchain-openai>=0.3.28
langchain-anthropic>=0.3.15
langchain-mcp-adapters>=0.1.6
langchain-deepseek>=0.1.2
langchain-tavily
langchain-groq>=0.2.4

# AI Model APIs
openai>=1.99.2
anthropic>=0.34.0

# Search and Research Tools
tavily-python>=0.5.0
arxiv>=2.1.3
pymupdf>=1.25.3
xmltodict>=0.14.2
linkup-sdk>=0.2.3
duckduckgo-search>=3.0.0
exa-py>=1.8.8
requests>=2.32.3
beautifulsoup4==4.13.3
markdownify>=0.11.6
aiohttp>=3.8.0

# Data Processing and Utilities
pydantic>=2.0.0
httpx>=0.24.0
rich>=13.0.0
pandas>=2.3.1

# Testing and Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Monitoring and Logging
langsmith>=0.3.37

# Cloud and Infrastructure
azure-identity>=1.21.0
azure-search>=1.0.0b2
azure-search-documents>=11.5.2
langchain-aws>=0.2.28

# Additional Tools
supabase>=2.15.3
mcp>=1.9.4